Add error handlers and refactor routes
