+++ dashboard/.env-example
@@ -1,0 +1,1 @@
+SERVER_URL=""
