--- dashboard/src/routes/users/+layout.server.ts
@@ -1,9 +1,0 @@
-import type { LayoutServerLoad } from './$types';
-
-export const load = (async ({ cookies }) => {
-	const token = cookies.get('token');
-
-	return {
-		token: token
-	};
-}) satisfies LayoutServerLoad;
