--- dashboard/src/routes/users/+layout.svelte
@@ -1,11 +1,0 @@
-<script lang="ts">
-	import type { PageData } from './$types';
-
-	export let data: PageData;
-</script>
-
-{#if data.token}
-	<slot />
-{:else}
-	<h1>Error</h1>
-{/if}
